+++ examples/deadlocks/deadlock_on_shutdown.py
@@ -6,6 +6,7 @@ This example highlights the fact that the ProcessPoolExecutor implementation
 from concurrent.futures is not robust to pickling error (at least in versions
 3.6 and lower).
 """
+
 import argparse