+++ tests/_executor_mixin.py
@@ -4,6 +4,7 @@ import math
 import pytest
 import threading
 from time import sleep
+import random
 
 from loky import TimeoutError, get_reusable_executor
 from loky.backend import get_context
@@ -18,8 +19,6 @@ except ImportError:
     psutil_exceptions = ()
 
 
-IS_PYPY = hasattr(sys, "pypy_version_info")
-
 # Set a large timeout as it should only be reached in case of deadlocks
 TIMEOUT = 40
 
@@ -30,6 +29,10 @@ def initializer_event(event):
     """Initializer that set a global test event for test synchronization"""
     global _test_event
     _test_event = event
+
+    # Inject some randomness in the initialization to reveal race conditions.
+    if random.random() < 0.2:
+        sleep(random.random() * 0.1)  # 0-100ms
 
 
 def _direct_children_with_cmdline(p):
@@ -148,9 +151,7 @@ def _check_executor_started(executor):
 
 
 class ExecutorMixin:
-    # PyPy is slower to create worker processes and creating
-    # too many processes can cause some tests to timeout.
-    worker_count = 2 if IS_PYPY else 5
+    worker_count = 5
 
     @classmethod
     def setup_class(cls):